+++ src/utils/qs.js
@@ -1,4 +1,9 @@
 /**
+ * Query string parser.
+ *
+ * To prevent HTTP Parameter Pollution only the last parameter will be
+ * serialized into the query record
+ *
  * @param {string} urlEncoded search parameters
  * @returns {Record<string,string>|{}}
  */
@@ -6,13 +11,7 @@ export function qs (urlEncoded) {
   const searchParams = new URLSearchParams(urlEncoded)
   const query = {}
   for (const [name, value] of searchParams.entries()) {
-    if (query[name]) {
-      Array.isArray(query[name])
-        ? query[name].push(value)
-        : (query[name] = [query[name], value])
-    } else {
-      query[name] = value
-    }
+    query[name] = value
   }
   return query
 }