+++ types/utils/qs.d.ts
@@ -1,4 +1,9 @@
 /**
+ * Query string parser.
+ *
+ * To prevent HTTP Parameter Pollution only the last parameter will be
+ * serialized into the query record
+ *
  * @param {string} urlEncoded search parameters
  * @returns {Record<string,string>|{}}
  */