+++ types/utils/safeServerShutdown.d.ts
@@ -3,16 +3,14 @@
  * alternative to [stoppable](https://github.com/hunterloftis/stoppable).
  * @param {Server} server the server instance
  * @param {object} [param1]
- * @param {number} [param1.gracefulTimeout=1000] graceful timeout for existing connections
- * @param {{info: function, error: function}} [param1.log] logger
+ * @param {number} [param1.gracefulTimeout=1000] (ms) graceful timeout for existing connections
+ * @param {Log} [param1.log] logger
  */
 export function safeServerShutdown(server: Server, param1?: {
     gracefulTimeout?: number | undefined;
-    log?: {
-        info: Function;
-        error: Function;
-    } | undefined;
+    log?: import("../../src/types").Log | undefined;
 } | undefined): void;
+export type Log = import('../../src/types').Log;
 export type HttpServer = import('http').Server;
 export type HttpSecureServer = import('https').Server;
 export type Http2Server = import('http2').Http2Server;